Prevent exposing mutating methods in SitePermissionMappings
