+++ src/main/java/org/craftercms/studio/api/v2/dal/security/RolePermissionMappings.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2007-2025 Crafter Software Corporation. All Rights Reserved.
+ * Copyright (C) 2007-2026 Crafter Software Corporation. All Rights Reserved.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 3 as published by
@@ -17,44 +17,12 @@
 package org.craftercms.studio.api.v2.dal.security;
 
 import java.util.Collection;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.Map;
 import java.util.Set;
-import java.util.regex.Pattern;
-import java.util.stream.Collectors;
-
-import static org.craftercms.studio.api.v2.security.ContentItemAvailableActionsConstants.mapPermissionsToContentItemAvailableActions;
-import org.craftercms.studio.permissions.StudioPermissionsConstants;
-import static org.craftercms.studio.permissions.StudioPermissionsConstants.SITE_WIDE_RULE_REGEXES;
 
 /**
- * Maps role rules to item available actions.
- * Instances will keep a map of rules to available actions for a given role.
- * It will also keep a set of site wide permissions. These are meant to be merged for all roles and then be used
- * to determine site-wide actions. e.g.: a user can be assigned a role with PUBLISH_REQUEST permission and another
- * one with PUBLISH_APPROVE permission. Such user would get the PUBLISH available action.
+ * Read-only view of role rules mapped to item available actions.
  */
-public class RolePermissionMappings {
-
-	// Rule path -> available actions
-	private final Map<Pattern, PermissionsActions> ruleContentItemPermissions = new HashMap<>();
-	private final Set<String> siteWidePermissions = new HashSet<>();
-
-	/**
-	 * Add a rule to this role mappings
-	 *
-	 * @param ruleRegex   regex to match the content item paths
-	 * @param permissions granted permissions for the rule
-	 */
-	public void addRuleContentItemPermissionsMapping(final String ruleRegex, final Collection<String> permissions) {
-		Pattern pattern = Pattern.compile(ruleRegex);
-		PermissionsActions permissionsActions = new PermissionsActions(permissions, mapPermissionsToContentItemAvailableActions(permissions));
-		ruleContentItemPermissions.put(pattern, permissionsActions);
-		if (SITE_WIDE_RULE_REGEXES.stream().anyMatch(ruleRegex::equals)) {
-			this.siteWidePermissions.addAll(permissions);
-		}
-	}
+public interface RolePermissionMappings {
 
 	/**
 	 * Get the available actions for a given path.
@@ -63,13 +31,7 @@ public class RolePermissionMappings {
 	 * @return available actions bitmap. This is calculated
 	 * by combining the available actions for all rules that match the path.
 	 */
-	public long getActionsForPath(final String path) {
-		return ruleContentItemPermissions.entrySet().stream()
-			.filter(entry -> entry.getKey().matcher(path).matches())
-			.map(Map.Entry::getValue)
-			.mapToLong(PermissionsActions::actions)
-			.reduce(0L, (a, b) -> a | b);
-	}
+	long getActionsForPath(String path);
 
 	/**
 	 * Get the permissions for a given path.
@@ -77,38 +39,20 @@ public class RolePermissionMappings {
 	 * @param path path of the content
 	 * @return set of permissions
 	 */
-	public Set<String> getPermissionsForPath(final String path) {
-		return ruleContentItemPermissions.entrySet().stream()
-			.filter(entry -> entry.getKey().matcher(path).matches())
-			.map(Map.Entry::getValue)
-			.map(PermissionsActions::permissions)
-			.flatMap(Collection::stream)
-			.collect(Collectors.toSet());
-	}
+	Set<String> getPermissionsForPath(String path);
 
 	/**
 	 * Get the site wide permissions for this role.
-	 * The site-wide permissions are the ones found in rules matching {@link StudioPermissionsConstants#SITE_WIDE_RULE_REGEXES}
 	 *
 	 * @return list of permissions
 	 */
-	public Collection<String> getSiteWidePermissions() {
-		return siteWidePermissions;
-	}
-
-	public Collection<String> getAllPermissions() {
-		return ruleContentItemPermissions.values().stream()
-			.map(PermissionsActions::permissions)
-			.flatMap(Collection::stream)
-			.collect(Collectors.toSet());
-	}
+	Collection<String> getSiteWidePermissions();
 
 	/**
-	 * Record to store permissions and actions for a given rule.
+	 * Get all permissions configured for this role.
 	 *
-	 * @param permissions permissions for the rule
-	 * @param actions mapped actions for the rule
+	 * @return collection of permissions
 	 */
-	protected record PermissionsActions(Collection<String> permissions, long actions) {
-	}
+	Collection<String> getAllPermissions();
+
 }
