+++ src/main/java/org/craftercms/studio/impl/v2/security/PermissionMappingsProviderImpl.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2007-2025 Crafter Software Corporation. All Rights Reserved.
+ * Copyright (C) 2007-2026 Crafter Software Corporation. All Rights Reserved.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 3 as published by
@@ -30,8 +30,10 @@ import org.craftercms.studio.api.v1.constant.StudioXmlConstants;
 import org.craftercms.studio.api.v1.exception.ServiceLayerException;
 import org.craftercms.studio.api.v2.dal.security.NormalizedGroup;
 import org.craftercms.studio.api.v2.dal.security.NormalizedRole;
-import org.craftercms.studio.api.v2.dal.security.RolePermissionMappings;
+import org.craftercms.studio.api.v2.dal.security.RolePermissionMappingsImpl;
 import org.craftercms.studio.api.v2.dal.security.SitePermissionMappings;
+import org.craftercms.studio.api.v2.dal.security.SitePermissionMappingsImpl;
+import org.craftercms.studio.api.v2.security.PermissionMappingsProvider;
 import org.craftercms.studio.api.v2.service.config.ConfigurationService;
 import org.craftercms.studio.api.v2.utils.StudioConfiguration;
 import static org.craftercms.studio.api.v2.utils.StudioConfiguration.CONFIGURATION_ENVIRONMENT_ACTIVE;
@@ -53,26 +55,28 @@ import com.google.common.cache.Cache;
  * Keeps a cached mapping of roles and available actions.
  * Allow access to path-constrained and site-wide available actions.
  */
-public class PermissionMappingsProvider {
+public class PermissionMappingsProviderImpl implements PermissionMappingsProvider {
 
-	private static final Logger logger = LoggerFactory.getLogger(PermissionMappingsProvider.class);
+	private static final Logger logger = LoggerFactory.getLogger(PermissionMappingsProviderImpl.class);
 
 	public static final String CACHE_KEY = ":available-actions";
 
-	private final Cache<String, SitePermissionMappings> cache;
+	private final Cache<String, SitePermissionMappingsImpl> cache;
 	private final StudioConfiguration studioConfiguration;
 	private final ConfigurationService configurationService;
 
-	public PermissionMappingsProvider(Cache<String, SitePermissionMappings> cache,
-									  StudioConfiguration studioConfiguration, ConfigurationService configurationService) {
+	public PermissionMappingsProviderImpl(Cache<String, SitePermissionMappingsImpl> cache,
+										  StudioConfiguration studioConfiguration,
+										  ConfigurationService configurationService) {
 		this.cache = cache;
 		this.studioConfiguration = studioConfiguration;
 		this.configurationService = configurationService;
 	}
 
+	@Override
 	public SitePermissionMappings getPermissionMappings(String site) throws ServiceLayerException {
 		var cacheKey = site + CACHE_KEY;
-		SitePermissionMappings mappings = cache.getIfPresent(cacheKey);
+		SitePermissionMappingsImpl mappings = cache.getIfPresent(cacheKey);
 		if (mappings == null) {
 			logger.debug("Cache miss for site '{}' cache key '{}'", site, cacheKey);
 			mappings = fetchSitePermissionMappings(site);
@@ -81,8 +85,8 @@ public class PermissionMappingsProvider {
 		return mappings;
 	}
 
-	private SitePermissionMappings fetchSitePermissionMappings(String site) throws ServiceLayerException {
-		SitePermissionMappings sitePermissionMappings = new SitePermissionMappings();
+	private SitePermissionMappingsImpl fetchSitePermissionMappings(String site) throws ServiceLayerException {
+		SitePermissionMappingsImpl sitePermissionMappings = new SitePermissionMappingsImpl();
 
 		String globalRolesConfigPath = studioConfiguration.getProperty(CONFIGURATION_GLOBAL_CONFIG_BASE_PATH) +
 			FILE_SEPARATOR + studioConfiguration.getProperty(CONFIGURATION_GLOBAL_ROLE_MAPPINGS_FILE_NAME);
@@ -111,7 +115,7 @@ public class PermissionMappingsProvider {
 		return sitePermissionMappings;
 	}
 
-	private void loadPermissions(Document document, SitePermissionMappings sitePermissionMappings) {
+	private void loadPermissions(Document document, SitePermissionMappingsImpl sitePermissionMappings) {
 		Element permissionsRoot = document.getRootElement();
 		if (permissionsRoot.getName().equals(StudioXmlConstants.DOCUMENT_PERMISSIONS)) {
 			Element siteNode = (Element) permissionsRoot.selectSingleNode(StudioXmlConstants.DOCUMENT_ELM_SITE);
@@ -122,7 +126,7 @@ public class PermissionMappingsProvider {
 			List<Node> roleNodes = permissionsRoot.selectNodes(StudioXmlConstants.DOCUMENT_ELM_PERMISSION_ROLE);
 			for (Node roleNode : roleNodes) {
 				String roleName = roleNode.valueOf(StudioXmlConstants.DOCUMENT_ATTR_NAME);
-				RolePermissionMappings rolePermissionMappings = new RolePermissionMappings();
+				RolePermissionMappingsImpl rolePermissionMappings = new RolePermissionMappingsImpl();
 				List<Node> ruleNodes = roleNode.selectNodes(StudioXmlConstants.DOCUMENT_ELM_PERMISSION_RULE);
 				ruleNodes.forEach(r -> {
 					String regex = r.valueOf(StudioXmlConstants.DOCUMENT_ATTR_REGEX);
@@ -138,7 +142,7 @@ public class PermissionMappingsProvider {
 		}
 	}
 
-	private void loadRoles(Document document, SitePermissionMappings sitePermissionMappings) {
+	private void loadRoles(Document document, SitePermissionMappingsImpl sitePermissionMappings) {
 		Element root = document.getRootElement();
 		if (root.getName().equals(StudioXmlConstants.DOCUMENT_ROLE_MAPPINGS)) {
 			Map<NormalizedGroup, List<NormalizedRole>> rolesMap = new HashMap<>();
